CHF-52 (CMN-204) Fix URI encoding/decoding issue
`URI` has 2 main constructors:
* a single argument one that expects an encoded value
* a multi-arguments one that expects decoded values

We were using one or the other (`Uris.create()` is using the single param
one), depending on the setter type on `MutableUri`:
* `setRawXxx()` are expecting encoded / raw values
* `setXxx()` are expecting encoded values

URI decoding is not a 2-ways operation: once the encoded chars have
been decoded, it's possible that they can't be encoded anymore.
For example `%2B` is an encoded value that becomes a `+` when decoded,
but `+` is a valid character in some components of the URI and may not
be encoded back into `%2B` as one would expect.

So, here, we're preserving the encoded URI format for all the methods
(even those that are accepting a decoded value).
The trick was to leave specific URI components decoding to the `URI` class
(hence the temporary `URI` instances in order to have the nice `getRaxXxxx()`
methods that give us properly encoded values for free).

Also added a bunch of unit tests